refactor: remove dead cloudAgent, createJob, localDockerRun, archiveService, and image fallback constants
