+++ server/app/src/types.ts
@@ -1,0 +1,4 @@
+export interface ActionData {
+  payload: { projectId: string; [key: string]: any }
+  [key: string]: any
+}
